requirements.txt
click==8.0.3
colorama==0.4.4
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap4==4.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pyotp==2.6.0
qrcode==7.3.1
visitor==0.1.3
Werkzeug==2.0.2
